Move all the global styles logic to the global styles folder
